ai-models/requirements.txt
# Python AI Backend Requirements
# Install with: pip install -r requirements.txt

# Core ML Libraries
ultralytics==8.0.196        # YOLOv8
xgboost==2.0.3              # Glucose prediction
scikit-learn==1.3.2         # ML utilities
numpy==1.24.3               # Numerical computing
pandas==2.0.3               # Data processing

# Computer Vision
opencv-python==4.8.1.78     # Image processing
pillow==10.1.0              # Image handling

# Web Framework
flask==3.0.0                # API server
flask-cors==4.0.0           # CORS support

# Utilities
python-multipart==0.0.6     # File uploads
joblib==1.3.2               # Model persistence
python-dotenv==1.0.0        # Environment variables
requests==2.31.0            # HTTP requests for LogMeal API
